# packages/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: MeiMart API
  version: "0.3"
  description: MeiMart 三端统一后端 API（契约 v0.2 基础 + v0.3 决策覆盖）
servers:
  - url: http://localhost:3000/api/v1
    description: dev
  - url: https://staging-api.meimart.xxx/api/v1
    description: staging
...
paths:
...
  /api/v1/admin/platform/audit-logs:
    get:
      tags:
        - platform
      description: 审计日志列表（按 user/resource/action/perspective/时间筛选，游标分页）
      parameters:
        - schema:
            type: string
            format: uuid
          required: false
          name: userId
          in: query
        - schema:
            type: string
          required: false
          name: resourceType
          in: query
        - schema:
            type: string
          required: false
          name: action
          in: query
        - schema:
            type: string
          required: false
          name: perspective
          in: query
        - schema:
            type: string
            format: date-time
          required: false
          name: from
          in: query
        - schema:
            type: string
            format: date-time
          required: false
          name: to
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          name: limit
          in: query
        - schema:
            type: string
          required: false
          name: cursor
          in: query
      responses:
        "200":
          description: 审计日志列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            userId:
                              type: string
                              nullable: true
                              format: uuid
                            action:
                              type: string
                            resourceType:
                              type: string
                            resourceId:
                              type: string
                              nullable: true
                            deviceType:
                              type: string
                              nullable: true
                              enum: *a8
                            perspective:
                              type: string
                              nullable: true
                            ip:
                              type: string
                              nullable: true
                            createdAt:
                              type: string
                              format: date-time
                          required:
                            - id
                            - userId
                            - action
                            - resourceType
                            - resourceId
                            - deviceType
                            - perspective
                            - ip
                            - createdAt
                      nextCursor:
                        type: string
                        nullable: true
                      hasMore:
                        type: boolean
                      total:
                        type: integer
                    required:
                      - items
                      - nextCursor
                      - hasMore
                required:
                  - success
                  - data
  /api/v1/admin/platform/audit-logs/{id}:
    get:
      tags:
        - platform
      description: 审计日志详情（含 before/after 快照）
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: 审计日志详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      userId:
                        type: string
                        nullable: true
                        format: uuid
                      action:
                        type: string
                      resourceType:
                        type: string
                      resourceId:
                        type: string
                        nullable: true
                      deviceType:
                        type: string
                        nullable: true
                        enum: *a8
                      perspective:
                        type: string
                        nullable: true
                      ip:
                        type: string
                        nullable: true
                      createdAt:
                        type: string
                        format: date-time
                      beforeData:
                        nullable: true
                      afterData:
                        nullable: true
                      userAgent:
                        type: string
                        nullable: true
                      traceId:
                        type: string
                        nullable: true
                    required:
                      - id
                      - userId
                      - action
                      - resourceType
                      - resourceId
                      - deviceType
                      - perspective
                      - ip
                      - createdAt
                      - userAgent
                      - traceId
                  message:
                    type: string
                required:
                  - success
                  - data
        "404":
          description: AUDIT_LOG_NOT_FOUND
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        pattern: ^E-(AUTH|COMMON|ORDER|PAYMENT|WAREHOUSE|USER|CATALOG|DISPATCH|RIDER|NOTIFY|PLATFORM|SETTLE|IM|AUDIT)-\d{3}$|^E-HTTP-\d{3}$
                      message:
                        type: string
                      details:
                        type: object
                        additionalProperties:
                          nullable: true
                    required:
                      - code
                      - message
                required:
                  - success
                  - error
  /api/v1/admin/platform/audit-logs/export:
    get:
      tags:
        - platform
      description: 审计日志导出 CSV（同 query 参数，最多 10000 行）
      parameters:
        - schema:
            type: string
            format: uuid
          required: false
          name: userId
          in: query
        - schema:
            type: string
          required: false
          name: resourceType
          in: query
        - schema:
            type: string
          required: false
          name: action
          in: query
        - schema:
            type: string
          required: false
          name: perspective
          in: query
        - schema:
            type: string
            format: date-time
          required: false
          name: from
          in: query
        - schema:
            type: string
            format: date-time
          required: false
          name: to
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          name: limit
          in: query
        - schema:
            type: string
          required: false
          name: cursor
          in: query
      responses:
        "200":
          description: CSV 流（text/csv）
...